# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Ultra Computer API
  version: 0.9.0
  description: |
    Full REST API for Ultra Computer — an AI orchestration platform with multi-agent
    swarm coordination, autonomous scheduling, voice, browser automation, marketplace,
    identity, messaging, and protocol interoperability (A2A, MCP, CLI, webhooks).
servers:
  - url: http://localhost:5000
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Bearer JWT token
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key authentication
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
      required: [ok]
...
    Message:
      type: object
      properties:
        id:
          type: string
        conversationId:
          type: string
        role:
          type: string
          enum: [user, assistant, system, tool]
        content:
          type: string
          nullable: true
        modelId:
          type: string
          nullable: true
        taskId:
          type: string
          nullable: true
        createdAt:
          type: integer
...
paths:
...
  /api/conversations/{id}/messages:
    get:
      operationId: getMessages
      summary: Get all messages for a conversation
      tags: [Conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Message'
    post:
      operationId: sendMessage
      summary: Send a message and trigger orchestration
      tags: [Conversations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content]
              properties:
                content:
                  type: string
                role:
                  type: string
                  default: user
                modelId:
                  type: string
      responses:
        '200':
          description: Created message
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
...
  /api/swarm:
    get:
      operationId: listSwarmLegacy
      summary: "[Legacy] List swarm sessions"
      tags: [Swarm]
      responses:
        '200':
          description: Sessions
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    post:
      operationId: createSwarmLegacy
      summary: "[Legacy] Create a swarm session"
      tags: [Swarm]
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
  /api/swarm/{id}:
    get:
      operationId: getSwarmLegacy
      summary: "[Legacy] Get swarm by ID"
      tags: [Swarm]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Swarm state
          content:
            application/json:
              schema:
                type: object
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      operationId: deleteSwarmLegacy
      summary: "[Legacy] Delete a swarm"
      tags: [Swarm]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'

  # ─── Autonomy ──────────────────────────────────────────────────────────────
...
  /api/cache/clear:
    post:
      operationId: clearCache
      summary: Clear cache (optionally by tier or modelId)
      tags: [Cache]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                tier:
                  type: string
                  enum: [exact, prefix, semantic]
                modelId:
                  type: string
      responses:
        '200':
          description: Cleared
          content:
            application/json:
              schema:
                type: object
...
  /api/cache/policy:
    post:
      operationId: setCachePolicy
      summary: Set a route-specific cache policy
      tags: [Cache]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [route, policy]
              properties:
                route:
                  type: string
                policy:
                  type: object
      responses:
        '200':
          description: Policy set
          content:
            application/json:
              schema:
                type: object
...
  /api/protocols/mcp/servers/{id}:
    delete:
      operationId: disconnectMCPServer
      summary: Disconnect / remove an MCP server connection
      tags: [Protocols]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Disconnected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...